+++ tests/testthat/test-plot-from-vignettes.R
@@ -2,6 +2,7 @@ skip_on_os(c("mac", "solaris"))
 skip_if_not_installed("ggplot2")
 skip_if_not_installed("datawizard")
 skip_if_not_installed("vdiffr")
+skip_on_cran()
 
 test_that("plot, vignette", {
   data(efc, package = "ggeffects")
@@ -50,12 +51,7 @@ test_that("plot, vignette", {
   # use 'one_plot = FALSE' for returning multiple single plots
   vdiffr::expect_doppelganger(
     "Vignette-plotintro-7",
-    plot(dat, one_plot = TRUE)
-  )
-
-  vdiffr::expect_doppelganger(
-    "Vignette-plotintro-8",
-    plot(dat, one_plot = TRUE, n_rows = 2) + ggplot2::theme(legend.position = "bottom")
+    plot(dat)
   )
 
   # dashed lines for CI
@@ -76,5 +72,49 @@ test_that("plot, vignette", {
   vdiffr::expect_doppelganger(
     "Vignette-plotintro-11",
     plot(dat, facets = TRUE, ci_style = "errorbar", dot_size = 1.5)
+  )
+})
+
+
+test_that("plot, vignette introduction", {
+  data(efc, package = "ggeffects")
+
+  # make categorical
+  efc <- datawizard::to_factor(efc, c("c161sex", "e42dep"))
+  # fit model with 4-way-interaction
+  fit <- lm(neg_c_7 ~ c12hour * barthtot * c161sex * c172code, data = efc)
+  # adjusted predictions for all 4 interaction terms
+  pr <- predict_response(fit, c("c12hour", "barthtot", "c161sex", "c172code"))
+  vdiffr::expect_doppelganger(
+    "Vignette-introduction-4-way",
+    plot(pr) + ggplot2::theme(legend.position = "bottom")
+  )
+  # fit model with 5-way-interaction
+  fit <- lm(neg_c_7 ~ c12hour * barthtot * c161sex * c172code * e42dep, data = efc)
+  # adjusted predictions for all 5 interaction terms
+  pr <- suppressWarnings(predict_response(fit, c("c12hour", "barthtot", "c161sex", "c172code", "e42dep"))) # nolint
+  vdiffr::expect_doppelganger(
+    "Vignette-introduction-5-way-1",
+    plot(pr)
+  )
+  vdiffr::expect_doppelganger(
+    "Vignette-introduction-5-way-2",
+    plot(pr, n_rows = 2) + ggplot2::theme(legend.position = "bottom")
+  )
+  vdiffr::expect_doppelganger(
+    "Vignette-introduction-5-way-3",
+    plot(pr, n_rows = 4) + ggplot2::theme(legend.position = "bottom")
+  )
+
+  # check one_plot
+  out <- plot(pr, one_plot = FALSE)
+  expect_length(out, 4)
+  vdiffr::expect_doppelganger(
+    "Vignette-introduction-5-way, single-1",
+    plot(out[[1]])
+  )
+  vdiffr::expect_doppelganger(
+    "Vignette-introduction-5-way, single-2",
+    plot(out[[4]])
   )
 })