+++ NEWS.md
@@ -7,8 +7,7 @@
   grid is used to plot all relevant panels. For five focal terms, multiple
   plots with facet grids are arranged using the **patchwork** package.
 
-* `values_at()` gains a new `"threenum"` option. `"quartiles"` and `"quartiles2"`
-  can be used as full-name alternatives to `"quart"` and `"quart2"`.
+* `values_at()` gains a new `"threenum"` option.
 
 # ggeffects 2.0.0